Split sync CRUD contract modules
